requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.4
openai==1.12.0
python-dotenv==1.0.0
dropbox==11.36.2
scikit-learn==1.3.2 